Bug Fixes & Change to powerup effect
Big line and sharp angles powerups triggered the hit detector. This is
now fixed.
The fast and slow powerup now use an exponential. Also the turn speed is
faster/slower for slow/fast powerup.

+++ classes/player.h
@@ -35,8 +35,8 @@ class Player{
         bool keyrelease;
         //
         int linewidth;
-        int shift;
-        int turn;
+        float shift;
+        float turn;
         bool rightangle;
         bool invisible;
         bool inverted;
